Added validators for character creation (not typing a number would cause program to crash), as well as validators for battle.
Added mana validator to cause an attack to fail if player does not have
enough mana to cast spell.

Added shop and adventure class to GameManager for future modes.

TO-DO: Display mana cost to players in battle, add functionality to spells (Lifesteal, etc), and add more enemies for players to choose to fight in battle (renamed Arena at some point - will be for boss battle practice).